+++ src/gfxlib2/win32/gfx_win32.c
@@ -688,11 +688,11 @@ void fb_hWin32SetMouse(int x, int y, int cursor, int clip)
 {
 	POINT point;
 
-	if (x != 0x80000000 || y != 0x80000000) {
-		if (x == 0x80000000) {
+	if (x != (int)0x80000000 || y != (int)0x80000000) {
+		if (x == (int)0x80000000) {
 			x = mouse_x;
 		}
-		else if (y == 0x80000000) {
+		else if (y == (int)0x80000000) {
 			y = mouse_y;
 		}
